Correlation Coefficient and Regression lines
1. Karl Pearson’s Correlation Coefficient:
Given a set of n observations 𝑥𝑖,𝑦𝑖 , i = 1,2,…,n of the pair of variables x and y. Then the Karl Pearson’s correlation coefficient for studying the degree of linear relationship between the two variables x and y is defined as 𝑟 𝑥,𝑦 =1𝑛 𝑥𝑖−𝑥 𝑦𝑖−𝑦 𝑛𝑖=1 1𝑛 𝑥𝑖−𝑥 2 1𝑛 𝑦𝑖−𝑦 2 where 𝑥 = 1𝑛 𝑥𝑖𝑛𝑖=1 and 𝑦 = 1𝑛 𝑦𝑖𝑛𝑖=1. Write an algorithm to determine Karl Pearson’s Correlation Coefficient.
2. Regression lines:
Given a set of n observations 𝑥𝑖,𝑦𝑖 , i = 1,2,…,n of the pair of variables x and y. If the variables in a bivariate distribution are related, then the scatter diagram of that distribution will concentrate around some curve called as curve of regression. If this curve is a straight line then the line is called line of regression. The line of regression is the line of best fit and is obtained by the principle of least squares. The equation of line of regression of y on x is: 𝑦−𝑦 = 𝑏𝑦𝑥 𝑥−𝑥 where 𝑏𝑦𝑥=𝑟𝜎𝑦𝜎𝑥 The equation of line of regression of x on y is: 𝑥−𝑥 = 𝑏𝑥𝑦 𝑦−𝑦 where 𝑏𝑥𝑦=𝑟𝜎𝑥𝜎𝑦 Here 𝜎𝑥= 1𝑛 𝑥𝑖−𝑥 2𝑛𝑖=1 and 𝜎𝑦= 1𝑛 𝑦𝑖−𝑦 2𝑛𝑖=1 are the standard deviations. Write an algorithm to determine regression lines.
3. Find the coefficient of correlation from the following data:
x
-10
-5
0
5
10
y
5
9
7
11
13
4. Find the two regression line from the following data:
x
1
2
3
4
5
6
7
8
9
10
y
10
12
16
28
25
36
41
49
40
50